requirements.txt
ccxt==4.2.18
pandas==2.1.4
numpy==1.26.2
ta==0.10.2
requests==2.31.0
python-dotenv==1.0.0